fx tweak

+++ data/sineShine.glsl
@@ -1,6 +1,6 @@
 #ifdef GL_ES
-precision mediump float;
-precision mediump int;
+precision highp float;
+precision highp int;
 #endif
  
 #define PROCESSING_TEXTURE_SHADER
